🚐 move types to own folder
