+++ api/src/api/v1/users/register/confirm.ts
@@ -15,7 +15,7 @@ import {
 } from '../../../../models';
 import * as QRCode from '../../../../services/qrcode';
 import * as PdfService from '../../../../services/pdf';
-import { RegisterConfirm } from '../../types';
+import { RegisterConfirm } from '../../types/types';
 import { RoleEnum } from '../../../../models/types';
 import Roles from '../../../../models/role';
 import { Tokens } from '../../../../models/token';
